feat (puconvocation-com): Add attendee table component
